Fixed Bugs
The last issue/bug is when I directly change the data in Firebase, it will cause duplication in the views since it automatically reload data from Firebase if it's changed.

Solution:
I modified and specify the condition for appending data to collections.
And specify the changes.

result:
The collection still reload data if it's changed. But it can recognize updates or adding new data.
when the view is reloaded, the new imformation will be dislplayed