Testing functionalities on my mobile phone
work as what I expect on my mobile phone.

TODO:
need to finish the functionality to display the interests with people.

Since I change the way of calling loading function, so in the Interest page does not need to reload the data from database again. It's finished by displaying the HomeView(root)